+++ src/main/java/com/dea42/watchlist/controller/RoamiotodoController.java
@@ -39,8 +39,8 @@ import lombok.extern.slf4j.Slf4j;
  * Copyright: Copyright (c) 2001-2021<br>
  * Company: RMRR<br>
  *
- * @author Gened by com.dea42.build.GenSpring version 0.7.1<br>
- * @version 0.7.1<br>
+ * @author Gened by com.dea42.build.GenSpring version 0.7.2<br>
+ * @version 0.7.2<br>
  */
 @Slf4j
 @Controller
@@ -74,7 +74,7 @@ public class RoamiotodoController {
 
 	@GetMapping("/new")
 	public ModelAndView showNewPage() {
-		return showEditPage(0);
+		return showEditPage(0l);
 	}
 
 	@PostMapping(value = "/search")
@@ -178,7 +178,7 @@ public class RoamiotodoController {
 	}
 
 	@GetMapping("/edit/{id}")
-	public ModelAndView showEditPage(@PathVariable(name = "id") Integer id) {
+	public ModelAndView showEditPage(@PathVariable(name = "id") Long id) {
 		ModelAndView mav = new ModelAndView("edit_roamiotodo");
 		Roamiotodo roamiotodo = null;
 		if (id > 0)
@@ -189,7 +189,7 @@ public class RoamiotodoController {
 	}
 
 	@GetMapping("/delete/{id}")
-	public String delete(@PathVariable(name = "id") Integer id) {
+	public String delete(@PathVariable(name = "id") Long id) {
 		roamiotodoService.delete(id);
 		return "redirect:/roamiotodos";
 	}
