+++ src/main/java/com/dea42/watchlist/repo/ShowsRepository.java
@@ -11,10 +11,10 @@ import com.dea42.watchlist.entity.Shows;
  * Copyright: Copyright (c) 2001-2021<br>
  * Company: RMRR<br>
  *
- * @author Gened by com.dea42.build.GenSpring version 0.7.1<br>
- * @version 0.7.1<br>
+ * @author Gened by com.dea42.build.GenSpring version 0.7.2<br>
+ * @version 0.7.2<br>
  */
 @Repository
-public interface ShowsRepository extends CrudRepository<Shows, Integer>,
+public interface ShowsRepository extends CrudRepository<Shows, Long>,
 JpaSpecificationExecutor<Shows> {
 }
