+++ src/main/java/com/dea42/watchlist/service/NetworksServices.java
@@ -32,8 +32,8 @@ import org.springframework.stereotype.Service;
  * Copyright: Copyright (c) 2001-2021<br>
  * Company: RMRR<br>
  *
- * @author Gened by com.dea42.build.GenSpring version 0.7.1<br>
- * @version 0.7.1<br>
+ * @author Gened by com.dea42.build.GenSpring version 0.7.2<br>
+ * @version 0.7.2<br>
  */
 @Slf4j
 @Service
@@ -150,11 +150,11 @@ public class NetworksServices {
 		return networksRepository.save(networks);
 	}
 	
-	public Networks get(Integer id) {
+	public Networks get(Long id) {
 		return networksRepository.findById(id).get();
 	}
 	
-	public void delete(Integer id) {
+	public void delete(Long id) {
 		networksRepository.deleteById(id);
 	}
